Add documentation for lambda-lifted representation
The code behind the "Lifted" intermediate representation used in Idris 2
compilation has been left mostly uncommented, until now. All features
marked as "public" have been given doc-strings, with sections on the
majority of arguments passed to functions.

A downside to having this much documentation is that the :doc REPL
command becomes quite verbose when used on e.g. Lifted: over a screen of
text is printed!

There didn't seem to be a consistent line wrap limit in this file, so
all code and documentation have been made to fit in a 80-character-wide
limit. Some function arguments were also given names, as a prerequisite
for writing about them in doc-strings.

There are a few TODOs added to this file: there are some parts of the
lambda-lifted code that I don't fully understand, and I'd appreciate
some advice on where I could look to fill in the gaps in my knowledge.
Generally, it's probably a good idea to have someone who understands
this code well to read through this new documentation, as I'm not
entirely confident in my own comprehension.