[ cleanup ] Remove extraneous show and simplify printLn
The primary goal of that commit is to remove extraneous `show` calls
while combined with `printLn` in some unit tests, which even though do
not hurt, are a bit confusing for the novice.

The simplifications of `print` and `printLn` are questionably useless
but I thought I'd include them too, for the fame perhaps.

Finally, I had to replace "5" by 5 in the expected result of test
basic048.  I believe this is the correct thing to do, unless the
creator of that test precisely wanted to test showing a string, which
seems surprising.