[ repl ] WIP: Put help text in knownCommands
This ensures that the help text is only in one place, avoiding the risk
of desync between the different texts, while still preserving that only
a short initial line of help is shown via `GenericHelp`.

I decided to go with `split ((==) '\n')` rather than `lines` since
`split` is guaranteed to return a non-empty list (meaning it'll return
the first line of text if there only is one), whereas `lines` is not.

TODO:
  [x] swap `simpleStr` for `cmdName`
  [x] merge `HelpHelp` with `Help` command
  [x] allow `:? :set` and friends
  [x] test cases for `:?` and `:let`
  [x] try moving all the help to `knownCommands` and use
      `split ((==) '\n')` for the short/generic help
  [ ] add detailed help for `:let` and `:help` itself
  [ ] tidy the naming