[ docs ] Build IdrisDoc in a namespace.
When invoking `idris2 --mkdocs foobar.ipkg`, the current behaviour is for the compiler to dump the documentation in `${builddir}/docs`.
The location `${builddir}/docs` is used regardless of the package name.
Thus for projects that present multiple packages the `docs` build directory will become polluted and corrupt if the developer is not careful.
This one change ensures that for package `x` the documentation is generated in the directory: `${builddir}/docs/x`.

Idris2 `Makefile` has been updated to reflect the change.