[ fix ] don't erase argument when type is erased
When the linearity is called on an application in erasure mode, the
function type containing the relevant quantity may already be erased. In
that case, we return the original term and an empty usage.

Returning an empty usage is potentially incorrect but no linearity is
done in erasure mode, making his safe.