Fixed spelling typo
Sorry about that!

(I noticed just now that there are a few uses of a bare `:exec` in some later chapters.  Maybe they all need to be followed by `main`, but I'm not sure. I haven't gotten that far in the book.  That's why I didn't just make a comment on chapter 1 that `:exec` with no argument should always be replaced by `:exec main`.)