Update typedd.rst: fixed typo in function name, added page ref
readNumber_v2 should be readNumbers_v2.  I also added a remark pointing to the corresponding function definition in the book, for people who prefer to type  in the definitions themselves (like me) rather than using the source files in the repo. 

There's another issue with the chapter 5 tips, but I'm not sure what the right fix is, so I'll submit an issue.