Update typedd.rst: Chapter 5: new imports for exercises in 5.3.4
The instructions for the exercises at the end of section 5.3.4 list several file-handling functions needed for the exercises.  They were in the Prelude, but now are in `System.File.Handle` and `System.File.ReadWrite`, and I added a note about that.  (`import System.File` would work for the exercise code, but the docs aren't there, so I didn't mention that.  I also didn't mention there a few other definitions, in other modules, that one needs to look up to understand the functions that the book lists.  I think that figuring out that one needs to find those additional docs is supposed to be part of the exercise.)