Refactor S-Exp protocols
Remove dependency on a lot of the compiler source-code.

This means specialising some of the S-Expr parser from the general
Idris2 source-code parsing code to a separate copy for the S-Expr parser.

It's difficult to share source-code between the two parsers because
the S-exp parser doesn't need the full token structure an Idris2
source-code requires.