[ base ] Add IEEE floating-point eps and ru
Racket provides a builtin for the epsilon `epsilon.0`. Chez Scheme does
not, but we can calculate it manually.

The (machine-)epsilon is the minimum distance between floating-point
numbers. The rounding unit is the smallest number which does not move a
float (iiuc), typically defined as eps/2 (or rather, eps is sometimes
2ru).