When searching for packages, rule out incompatible TTC versions
Prior to this change, the first package matching version requirements
would be used by the compiler -- if that package was only installed with
an incompatible TTC version, it simply cannot be used. After this
change, we skip package directories with only incompatible TTC versions
which allows the compiler to find a farther down option (if one exists)
in some other directory that does have a compatible TTC vesrion.