Add WithData and Record
The diff is huge but that is exactly why this change is needed. Changing
anything in the Idris' tree, wheteher it's Syntax or RawImp is a
nightmare. Each change requires:
- updating all functions using them (fine)
- updating all the printing functions
- updating all the serialising functions
- updating all the reflection functions
- updating all traversal functions
- implementing Functor and traverse for Core

`Record` allows to label products of types and `WithData` attaches a
record to a piece of data, such that the existing data can still
be used normally even after we add data to it.

This change should also remove the need for types such as `WithFC` and
`WithName`.