[ feature ] Add `CITATION.cff`
GitHuB supports the
[CFF](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files)
format for displaying how the software *should* be cited.

Let us add a note on how Idris2 should be cited.

We add a preference for the Idris2 paper as the preferred citation

GitHub also [supports pointing to a BibTex
file](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files#other-citation-files).