[docs] Group mailing list and discord
The website is not really related to questions, so it makes more sense
to group the mailing list with the discord.  As part of that separation,
then, add a brief sentence on what the website does.

Similar to the website, mention that the discord is currently the most
active interactive discussion place.