[ fix #399 ] Thread-safe way to invoke GC at program end (#3643)
According to Chez's docs, "the thread that invokes collect must be the
only active thread".  This implies that the only safe way to manually
trigger GC in a (potentially) multi-threaded program is by
`(collect-rendezvous)`.  That function will pause all threads and make
one of them call the `collect-request-handler`.

For this final GC, `collect-request-handler` is changed to a basic one
that forces a full collection followed by `(blodwen-run-finalisers)`.