[ v0.8.0-rc1 ] Generate bootstrap code
With the bootstrap code generated, bootstrapped, and regenerated using
the new compiler, this completes the preparation for a release
candidate!

The CHECKLIST will be cleared once the community is happy to go ahead
with the release (pending further RCs if adjustments are needed),
ideally after the tags, website tarball, etc. have been completed.