[ v0.8.0-rc2 ] Fix prefix in Racket bootstrap code
I had accidentally left in my local prefix in the racket code!  This
should be "__PREFIX__" (as per the CHECKLIST) in order for bootstrapping
to actually work.

Also reordered and clarified the `mkdist.sh` step in the CHECKLIST,
since this requires the tag on git to exist in order for the script to
bundle the right version of the code.