[ admin ] Add pack-db to release checklist
This caught me out when preparing the v0.8.0 release and had the CI fail
on the PR: there is a file in pack's collections directory which has the
compiler version pinned to a release version.  This needs to be updated
once the release and tag have been created on GitHub.  It turns out we
ran into this with v0.7.0 as well, so it seems worth retaining for the
future (at least until we figure out a hopefully automated solution).