[ nix ] Point emacs-src to idris-community/idris2-mode
This previously pointed at a custom fork (redfish64/idris2-mode).
However, based on GitHub's comparison of the two repos, the fork is one
commit behind and 0 ahead of the version on idris-community, so I
believe they should be interchangeable.  I'm switching the URL to point
to the community version, since if we're referencing it from the Idris2
repo, it should be something we can easily update and maintain.